extracts setupPreloadProtection static
