--- src/modules/protection/InspectorProtectionModule.ts
@@ -1,45 +1,0 @@
-/*
- * Copyright 2018-2020 DITA (AM Consulting LLC)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Developed on behalf of: DITA, Bokbasen AS (https://www.bokbasen.no), Bluefire Productions, LLC (https://www.bluefirereader.com/)
- * Licensed to: Bluefire Productions, LLC, Bibliotheca LLC, Bokbasen AS and CAST under one or more contributor license agreements.
- */
-
-import { addListener, launch } from 'devtools-detector';
-
-export interface InspectorProtectionModuleConfig {
-  clearOnInspect: boolean;
-  api: { inspectDetected?: () => void };
-}
-
-export default class InspectorProtectionModule {
-  public static start(config: InspectorProtectionModuleConfig): void {
-
-    const onInspectorOpened = (): void => {
-      if (config.clearOnInspect) {
-        console.clear();
-        window.localStorage.clear();
-        window.sessionStorage.clear();
-        window.location.replace(window.location.origin);
-      } else if (typeof config.api?.inspectDetected === "function") {
-        config.api.inspectDetected();
-      }
-    };
-
-    addListener(onInspectorOpened);
-
-    launch();
-  }
-}
